# backend/cmd/doc-service/docs/swagger.json
{
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "Document creation and listing APIs.",
    "title": "SyncDocs Document Service",
    "contact": {},
    "version": "1.0"
  },
  "basePath": "/",
  "paths": {
    "/docs": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "produces": ["application/json"],
        "tags": ["documents"],
        "summary": "List documents",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Document"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["documents"],
        "summary": "Create a document",
        "parameters": [
          {
            "description": "Document payload",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/handlers.CreateDocumentRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/models.Document"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/handlers.ErrorResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "handlers.CreateDocumentRequest": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "My first doc"
        }
      }
    },
    "handlers.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "models.Document": {
      "type": "object",
      "properties": {
        "created_at": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "owner_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Use \"Bearer \u003ctoken\u003e\"",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}